kitchenpos-web/src/main/java/kitchenpos/application/product/ProductService.java
package kitchenpos.application.product;

import java.util.List;
import kitchenpos.application.product.dto.ProductCreateRequest;
import kitchenpos.application.product.dto.ProductResponse;
import kitchenpos.application.product.dto.ProductsResponse;
import kitchenpos.domain.product.Product;
import kitchenpos.domain.product.ProductRepository;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@Service
public class ProductService {
    private final ProductRepository productRepository;

    public ProductService(final ProductRepository productRepository) {
        this.productRepository = productRepository;
    }

    @Transactional
    public ProductResponse create(final ProductCreateRequest request) {
        final Product product = new Product(request.getName(), request.getPrice());
        final Product savedProduct = productRepository.save(product);
        return ProductResponse.from(savedProduct);
    }

    public ProductsResponse list() {
        List<Product> products = productRepository.findAll();
        return ProductsResponse.from(products);
    }
}
